% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rfPredVar.R
\name{rfPredVar}
\alias{rfPredVar}
\title{rfPredVar}
\usage{
rfPredVar(rf, pred.data = NULL, CI = FALSE)
}
\arguments{
\item{rf}{A random forest trained with \code{keep.inbag=TRUE}. See details for more information.}

\item{pred.data}{The data used to \code{predict} with the \code{randomForest}.}

\item{CI}{Should 95\% confidence intervals be returned along with predictions and prediction variances?}
}
\value{
A data frame with the predictions and prediction variances (and optionally 95% confidence interval)
}
\description{
Generate predictions and prediction variances from a random forest based on the infinitesimal jackknife.
}
\details{
The original version of \code{randomForest} with the \code{keep.inbag=TRUE}
only keeps track if each training data point was or was not included in each resample.
Install a tweaked version of \code{randomForest} that amends this to include the number of times each
training data point was included in each resample by running the following code in R:
\code{devtools::install_github('cole-brokamp/randomForest')}
}
\examples{
library(randomForest)
data(airquality)
d <- na.omit(airquality)
rf <- randomForest(Ozone ~ .,data=d,keep.inbag=T,sampsize=30,replace=FALSE,ntree=500)
rfPredVar(rf,pred.data=d,CI=TRUE)
}

